mobile-store/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
reportlab
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
psycopg2-binary
